Controll led with UART

+++ local_inc/Output_task.h
@@ -11,7 +11,7 @@ int setup_Output_Task(int prio, xdc_String name, uint32_t wait_ticks);
 
 int setup_spi();
 
-int show_leds(uint8_t l);
+int show_leds(uint32_t l);
 
 int set_pwm(uint8_t pwm);
